+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Eclipse m2e - Release notes
 
+## 2.9.2
+
+### Multi-Release-Jar support
+
+JDT added support for native [https://openjdk.org/jeps/238](Multi-Release) compilation support recently.
+With that present, m2e can now automatically support the [maven-compiler-plugin](https://maven.apache.org/plugins/maven-compiler-plugin/) option [https://maven.apache.org/plugins/maven-compiler-plugin/compile-mojo.html#multiReleaseOutput](`multiReleaseOutput`)
+and configure such projects accordingly.
+
 ## 2.9.1
 
 * 📅 Release Date: 02nd September 2025